feat(app): init commit for 3.0
