+++ App.js
@@ -2,6 +2,7 @@ import { Provider } from 'react-redux';
 import { OrientationLocker, PORTRAIT } from 'react-native-orientation-locker';
 import { store } from './src/store/redux/store';
 import AppStatusBar from './src/components/status-bar';
+// import AppContent from './src/app-content/index_old';
 import AppContent from './src/app-content';
 import './src/utils/locales/i18n';
